Separated the friction on the FoosBar into two different components.  This is probably technically incorrect, I'm not totally sure if the rotational friction should work the same way as sliding friction.  In any case, it allows for more customization for better game play.  The coefficient for rotational "sliding" friction is now separated from the coefficient for linear sliding friction.
Changed some debug drawing and moved the boosters around a bit.  Added a bunch of variables to PlayState that determine positioning of everything.  Basically, there is a bounding box for the screen and field and a goal size, everything else is positioned and sized based on that.  Physics boundaries, debug drawing of the grass and stadium, etc.